+++ client/src/components/grants/New.tsx
@@ -4,7 +4,7 @@ import Button, { ButtonVariants } from "../base/Button";
 import ProjectForm from "../base/ProjectForm";
 import Cross from "../icons/Cross";
 import ExitModal from "../base/ExitModal";
-import VerificationForm from "../VerificationForm";
+import VerificationForm from "../base/VerificationForm";
 import { ProjectFormStatus } from "../../types";
 
 function NewProject() {
